requirements.txt
asgiref==3.10.0
bleach==6.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==5.2.7
django-appconf==1.2.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.2
django-summernote==0.8.20.0
django-taggit==6.1.0
django_compressor==4.6.0
fonttools==4.60.1
graphviz==0.21
idna==3.11
kiwisolver==1.4.9
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
pydotplus==2.0.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
rcssmin==1.2.2
requests==2.32.5
rjsmin==1.2.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.11.0
